Partially-working firmware with LoRa MAC v4.6.0
This patch makes a bigger change to lrw.[ch] and cmd.[ch] initiated by
the migration to LoRaMAC-node v4.6.0 with state management handled
through the library. This is a working firwmare version that works in
both OTAA and APB activation modes with TTN in the US915 region.

Partial list of changes:
- Re-design of the AT command interface (will eventually emulate Murata
  firmware API).
- NVM state management through LoRaMAC-node with direct write to EEPROM
  (without the config API)
- Flatter source code structure without unnecessary callback
  hierarchies.
- Smaller API in the lrw.c module. Let the AT API configure the MAC
  through the MIB which provides a much more general API.
- Support for selecting MAC version and ABP MAC version in the Makefile
  (only 1.0.4 or 1.1.1 can be selected for OTAA and 1.0.3 can be selected
  for ABP).
- Implemented region switching with all regions compiled into the
  firmware by default.