Store and modify active region in NVM data
This patch modifies the mechanism through which the active region is
selected and managed. We no longer keep the active region in an external
configuration structure. Instead, we modifiy the value kep in MacRegion2
and trigger NVM store when necessary. Since the active region needs to
be passed to the loramac initialization function, we also need to
implement a function to fetch the value before the internal NVM context
is restored.

The default active region, i.e., the region that will be automatically
activated if there is no NVM data can be configured in the Makefile.