Create a separate module for NVM and sys config data
The module encapsulates everything related to NVM (EEPROM). It
partitions the EEPROM memory and creates a partition for system
configuration data. The module also provides an API to access and
synchronize (store in EEPROM) system configuration data.

The system configuration data is meant for platform-related parameters
such as UART baudrate and timeout, payload format, etc.