Add support for composite log lines
This patch extends the logging API to make it possible to compose a
single log line via repeated invocations of the various log_* functions.
This is useful, e.g., if the caller needs to programmatically enable or
disable portions of the log string at runtime.

Usage example:
  log_compose();
  log_debug("a");
  log_debug(" b");
  log_debug(" c");
  log_finish();

The above code will produce a single log message consisting of "a b c".