Implement the command AT+LNCHECK
In the original firmware, the command takes an optional parameter of the
form AT+LNCHECK[=<0,1>]. When set to one, the link check operation will
be piggy-backed on the next uplink. When set to 0, the command will
internally send an empty uplink immediately.

Since our AT command parser cannot support the above syntax for commands
that are not getters/setters, we use the syntax AT+LNCHECK[ <0,1>]
instead.