Make AT+VER backwards compatible and add AT$VER
This patch changes the string returned by AT+VER? to be compatible with
the string returned by the original firmware. We modify the format of
the version string as well as the formware of build date. The version
string to be returned can be configured in Makefile.

Add the command AT$VER which returns a version string automatically
generated from the git repository. This is the original behavior of
AT+VER, it was just moved to AT$VER.

Examples:

AT+VER?
+OK=1.1.06,Mar 08 2022 03:00:54

AT$VER?
+OK=v1.0.0-rc1-1-gcfa05414 [LoRaMac v4.6.0],2022-Mar-08 03:00:54 EST