Implement DMA-based receive path on LPUART1
Before, the LPUART1 port was configured to generate an IRQ for each
received character. This approach did not work reliably during EEPROM
write activity intervals. EEPROM writes block Flash program reads,
resulting in lost IRQs during periods of heavy EEPROM write activity.

This patch configures the LPUART1 port to receive data via a combination
of DMA and IRQ. This approach is inspired by Tilen Majerle's examples
over at https://github.com/MaJerle/stm32-usart-uart-dma-rx-tx

We setup channel 6 on the DMA controller to copy data from the LPUART1
peripheral to a memory buffer. The channel is configured in a circular
mode, i.e., it wraps over to the beginning of the buffer once the end
has been reachead. We use a combination of DMA and UART IRQs to copy the
data from the buffer to the console FIFO. The DMA controller is
configured to generate an IRQ each time one half of the buffer has been
filled with data. The LPUART1 peripheral is configured to generate an
IRQ when the UART port becomes idle. We copy data from the DMA buffer to
the console FIFO in each of the IRQs.

Since the DMA controller is disabled in Stop mode, we additionally need
to switch LPUART1 back to the IRQ method before entering Sleep. We pause
DMA and reconfigure LPUART1 to wake up the core once a start bit has
been received. Once woken up, we enable DMA and prohibit the core from
entering Sleep until a "line idle" IRQ has been received from LPUART1.
This effectively ensures that DMA is running from the time a start bit
is received until the core is about to enter the Stop mode again.

Closes #7