Implement payload read timeout via AT+TO
When reading the payload for an uplink message, wait at most the number
of milliseconds configured via AT+TO. If the upload takes longer, we
terminate the read process and send a message with an incomplete payaload
(all bytes received before the timeout) to match the behavior of the
original firmware.

On formatting errors (with AT+DFORMAT=1), return an error to the client.

Closes #8