Review and improve OTAA and ABP activations
We no longer keep track of the currently active mode in a variable local
to the lrw.c module. Instead, we derive the activation mode from the
MIB. That way, we have a single authoritative source of that
information.

This patch also refactors how debugging information is printed during
LoRaMac initialization. We print the basic information about the device,
as well as the information about the current active network session
(if one exists).

Closes #9