Fix a potential race condition in console_flush
Before entering sleep, we need to disable IRQs and check that the
vcom_ready flag stil indicates that the console is not ready. Only if it
is not ready we go to sleep. This guarantees that sleep is never entered
after the IRQ which set the flag was triggered (this would potentially
make the CPU sleep forever).

Closes #18