Generate dependency files with system headers
The libraries under lib/ are treated as system libraries for the purpose
of header sanity checking in gcc. That would also make gcc with the
command line option -MMD exlude them from dependency files. Since the
files under lib/ are part of the application and can be modified, let's
include them in dependency files as well so that the binary gets
properly rebuilt whenever a header file under lib/ changes.

Closes #32