Add a compile-time option to restore chmask after Join
For deploymens with an incorrectly configured LoRaWAN network server
that pushes the wrong channel mask to the node, the compile-time option
RESTORE_CHMASK_AFTER_JOIN can be used to instruct the MAC to ignore the
channel mask pushed by the network server in Join Accept.

If enabled in the Makefile, the MAC will save the channel mask prior to
sending a Join. Upon receiving a Join Accept, the previously saved
channel mask is restored.

Closes #20