Make log level runtime-configurable
Use the new command AT$LOGLEVEL to query or set the log level on the
USART port at runtime. When compiled in release mode, all logging is
disabled by default. When compiled in debug mode, all logging is enabled
by default.

The parameter configured via AT$LOGLEVEL is NOT saved to the NVM and
will revert to the default value upon reboot.