Reorganize system initialization
Initialize the LPUART1 port for the AT command interface early before
all other subsystems so that we catch AT commands transmitted by the
host early upon reboot.

Remove the one second delay that was enforced at the beginning of the
main function in DEBUG mode. The goal of that delay was to make it
possible for the developer to see early debugging output on the USART
interface. Without the delay, that output will be lost of the
application isn't fast enough to read from the port upon flashing the
firmware. To see such early debugging output, issue a AT+REBOOT.