Make AT command parser more robust
This change improves the robustness of the AT command parser against
noise or arbitrary (non-AT) data. The new version can detect the AT
prefix in the middle of arbitrary input, e.g., the data generated when
the port is initially configured with the wrong baud rate.

This is more in-line with the original algorithm from the Hayes AT
command interface which detects AT command sequences in arbitrary data
stream. The Hayes approach uses a guard interface to only detect AT
commands after a short period of inactivity. We do not implement this
here since our AT command interface only supports AT commands, not
arbitrary data.