Clear UART error flags before delegating to HAL
We use the LPUAR1 port in DMA receive mode. On error, the HAL aborts the
DMA transfer. Since we don't want that, we clear the various error flags
before delegating to the HAL IRQ handler.

This was triggered when the port was configured with the wrong baud
rate. In that case, the LPUART1 peripheral sets the framing error flag.