Fix AT+FRMCNT
The counters are uint32_t and thus need to be printed as %lu.
Furthermore, we need to select either AFCntDown or FCntDown depending on
the LoRaWAN protocol version negotiated between the device and the
network server.

Note: UINT32_MAX is a special value that indicates that the counter is
set to the initial value, i.e., no uplink or downlink has taken place so
far.