Yield when saving state to NVM
In some cases, e.g., after an OTAA Join, pretty much all the state needs
to be written into NVM (EEPROM) which can take a while. To prevent the
starvation of the AT command interface, we write NVM groups individually
and yield after each write to give the AT command interface a chance to
process incoming commands.

Since there are no threads or processes in the modem, this is
implemented by preventing the system from going to sleep as long as
there is something to be written to NVM, and returning from the function
upon each NVM group write. The main loop will keep invoking the NVM
write function as long as there is at least one group that needs to be
written.

Closes #53