Redesign UART buffering and AT command processing
This patch brings in three interrelated improvements motivated by #50:
  1) AT command processor resistant to UART overruns
  2) Single-copy UART data buffering architecture
  3) Faster vectorized FIFO (queue) implementation

The changes in this commit are primarily motivated by #50. Since the AT
command interface does not check the return values of the various atci_*
functions that send data to the client, we need to make those blocking
so that they really send all data to the client. The MCU is put to sleep
if the transmission buffer is empty. Upon waking up, the function will
continue sending data until all data is sent. This change is primarily
motivated by class B and class C applications where the network may send
more data to the application and there would be risk of losing the data
on LPUART overruns.

In order to streamline the buffering on LPUART and USART ports, this
patch also redesignes the transmission and reception buffers (FIFOs).
The new implementation has a more efficient API and makes it possible to
transmit and receive all UART data with only a single memory copy per
byte. That is, bytes received over LPUART are copied once from the DMA
buffer to the FIFO buffer and processed by the ATCI directly from the
memory buffer backing the FIFO. Data transmitted by the application are
copied once into the transmission FIFO and the DMA controller is
programmed to transmit directly from the FIFO memory buffer.

Finally, this commit also replaces the original FIFO implementation
found in fifo.{c,h} with a much more efficient vectorized version which
uses memcpy to copy data. The new implementation, which can be found in
cbuf.{c,h} also has a much more flexible API designed allow the
application to minimize the time spent in a critical section (with IRQs
blocked). The API functions that need to be invoked in a critical
section have all constant running time.

The old wrapper code in console.{c,h} is also gone. The buffers and FIFO
data structures have been moved to lpuart.{c,h}, similar to how it was
implemented in usart.{c,h}. The LPUART TX and RX FIFOs are made public
to make it possible for the application to access those directly in
critical sections.

Closes #50