Perform reset only when there is no work to be done
When the application schedules a reset, we need to postpone the reset
until the main loop (thread) has no more work to do to give ongoing
tasks, e.g., NVM update, a chance to finish. The main thread has no more
work to when the application has no sleep lock active.

Fixes a bug introduced by the fix for #53