Preserve DevNonce across factory resets
In order to make it possible for the device to send an OTAA Join after a
factory reset, this patch modifies the AT+FACNEW procedure to preserve
the DevNonce value across factory resets. DevNonce is a counter used in
Joins that is meant to be stored persistently by the device.

Prior to this change, DevNonce was reset to 0 during factory reset.
Since the network server remembers the most recent DevNonce for a
device, this would cause Joins after factory reset to fail, until the
device is re-added to the network, or until it reaches a high enough
DevNonce value through Join retransmissions.

If you wish to reset the DevNonce during factory reset, use the
following form: AT+FACNEW=1

Closes #65