Implement randomized Join retransmissions
This patch changes the behavior of AT+JOIN. By default, the modem will
retransmit the Join request up to eight times if no answer is received
from the network server. Thus, a single AT+JOIN command can trigger up
to nine Join transmissions in total. This also means that it will take
longer to receive an +EVENT1,{0,1}, up to a minute if no Join answer is
received from the network.

The default number of Join transmissions is set to nine to make sure
that the Join is transmitted at least once in each eight-channel
subband in regions that use all 64 channels such as US915, plus one
extra transmission in the 500 kHz channel subband.

The application can control the maximum number of Join retransmissions
via an optional parameter to AT+JOIN=<retransmissions>. To disable
retransmissions use AT+JOIN=0. To only retransmit once use AT+JOIN=1,
etc. The maximum number of retransmissions that can be configured via
the optional parameter is 15.

The modem applies a small randomly generated delay before each
retransmission in accordance with Section 7 of LoRaWAN Specification
v1.1. The random number generator is seeded from the LoRa radio which
generates a random seed on each device. The random delay is between 100
ms and 500 ms.

Closes #61