Implement command AT$SESSION
Use this command to obtain the current network association information
such as activation mode, negotiated LoRaWAN protocol version, network
id, and device id. The command returns a list of comma-separated values:

AT$SESSION?
OK=public,OTAA,1.1.1,00000013,260C5A7F

Where the first value is either "public" or "private", the second value
represents the LoRaWAN activation mode and can be one of "None", "OTAA",
"ABP", "?", the third value represents the LoRaWAN protocol version
being used, the fourth value is the ID of the network and the fifth
value is the address of the device (DevAddr). The third through fifth
values will only be present if the activation mode is not "None".