USART1 clean up
The USART1 port is used in debugging mode for logging. The previous
implementation did not manage USART1 interrupts properly which could
have lead to deadlock in a couple of corner cases.

This version performs port initialization via STM HAL LL and also cleans
up the interrupt handler to make it simpler and more robust.