Build debugging functionality conditionally
This patch modifies the source code and the build system to only build
debugging functionality when the firmware is being built in debugging
mode. This functionality includes: USART1 port related code, logging
subsystem, HAL low-level drivers ll_usart and ll_rcc, and the Segger
debugging library.

When compiled in release mode, the above-mentioned modules will be
skipped. When compiled in debugging mode, the modules are included.

This change also makes the AT command AT$LOGLEVEL available in the
debugging version of the firmware only.