Perform Join retransmissions from the main execution thread
OTAA Join retransmissions are performed from within a timer handler.
Since the timer manager library invokes timer callbacks from the ISR
context, we modify the retransmission callback to simply set an event to
be processed from the main thread context on the next run of the main
function within the lrw.c module. This makes it possible to invoke
blocking functions from within the retransmission handler, e.g., to send
events to the application over ATCI.

Closes #70