More reliable halt function implementation
The halt function is invoked by the LoRa modem whenever it encounters
an irrecoverable problem, e.g., faulty hardware or critical software
error. The purpose of the halt function is to bring the modem into a
low-power state until it is reset via the external reset pin. The only
way to recover from the halted state should be via the external reset
pin or a power cycle. Ideally, the modem would not drain the device's
batter while it is being halted.

This patch improves the halt function to minimize the probability that
it will be spinning while being halted. We clear all sleep locks and
mask all EXTI interrupts and events before entering the low-power Stop
mode.

Closes #69