Cleanups in oscillator and clock configuration
This patch refactors the code that initializes the various oscillators
and clocks with the goal to make the code more straighforward and easier
to follow.

Also, disable a couple of components that were previously left enabled:
  * MSI is disabled upon switching system clock to PLL(HSI16)
  * LSI is disabled upon switching RTC clock to LSE
  * PWR clock is left disabled and is only enabled as needed

Furthermore, we also reconfigure the MCU to wake up from the Stop mode
with HSI16 enabled instead of MSI. This elimiates the need to switch
from MSI to HSI16 upon wakeup.

Closes #63