Add support for hard reboot
Without any parameters, the AT command AT+REBOOT performs a clean
reboot, where the modem waits for all active tasks to finish and then it
performs the reboot. The reboot is postponed until the modem is ready to
enter the low-power Stop mode, which serves as an indication of all
tasks being idle.

This patch implements support for hard reboot via 'AT+REBOOT 1'. In this
case, the modem will not wait for anything and will trigger a system
reboot immediatelly. Note that the application does not even get +OK in
this case.

'AT+REBOOT 0' performs clean reboot (it's an alias for 'AT+REBOOT')