New Python library for the TypeABZ modem with open firmware
The Python library can be either imported as a module into a larger
project, or it can be used from the command line to manage TypeABZ
Murata modem with the open firmware.

The command line interface can be used to obtain modem information,
configure the modem, manage security keys, and perform LoRaWAN network
operations. The command line tool is installed under the name lora.

The command line tool can only be used with the open firmware. The
Python module, however, provides a class called MurataModem designed to
work with the AT command interface provided by the original Murata Modem
firmware.