Do not reset DevEUI during factory reset by default
When AT+FACNEW is invoked without any arguments, DevEUI and DevNonce
values will be preserved across the reset to factory defaults. To reset
either of those values, pass an optional argument "flags" to AT+FACNEW.
When bit 0 is set to 1, DevNonce will be reset. When bit 1 is set to 1,
DevEUI will be reset. Thus, to reset everything, invoke AT+FACNEW 3.

Closes #79