Fix bug in get_buffer_from_hex
Previously, the function assumed that the value being parsed stretches
all the way to the end of the AT command value. This is not necessarily
always the case, especially in AT commands that accept a list delimited
with ,. We modify the function to accept one more parameter which can be
used to specify the maximum length of the value. If set to 0, the
function works as before.

Closes #83