More efficient build upon git version change
The previous version of the Makefile would re-build the entire project
if the version strings generated from the git repository changed. The
new version will only rebuild the files that really depend on these
strings. These are currently src/cmd.c and src/main.c.

To ensure that the Makefile is updated if another file is modified to
depend on these strings, we only include the preprocessor macros in the
two selected files. If a new file is made to depend on, say, VERSION,
the compilation will fail until a new target is added for the file which
properly depends on the strings generated from the git repository.