Update firmware and library version strings
The version strings for the firmware and for the LoRaMac-node library
are generated from git tags. Conventionally, the tag name is prefixed
with "v". However, the "v" prefix is not part of the version string and
we thus strip it if the tag name matches a three-component version
pattern.

Also, update the comment that incorrectly mentioned that the generated
version files are included as dependencies for all targets. That is no
longer the case.