Implement AT+RFPARAM
The getter variant of this command can be used to obtain a list of all
currently active radio channels. For each channel the command returns
the channel's center frequency and minimum and maximum data rates.

The setter variant of this command can be used add or remove RF
channels. To add a channel, pick a number that is currently not used and
provide the desired center frequency and data rates. To delete a
channel, use it's id and set the frequency and data rates to zero.

Closes #55