Implement AT+RSSITH and AT+CST
The AT+RSSITH command can be used to control the RSSI free channel
threshold parameter used by the listen before talk (LBT) algorithm in
the KR920 and AS923 bands. The value is in dBm.

The AT+CST command can be used to control the carrier sense time (CST)
parameter used by the LBT algorithm in the KR920 and AS923 bands. The
value is in milliseconds.

When used in a band that does not support LBT, these commands return
+ERR=-17.

Closes #57, closes #58