Remove conditional variable initialization
When this Makefile variable is set to any value, the functionality will
be enabled. Thus, do not make it possible for the application to
override the value, since it won't be handled correctly, e.g., value 0
to disable it won't work. The correct way to disable the functionality
is to comment out the Makefile variable.