Make OTAA Join data rate configurable
The LoRaWAN community recommends random varying of the data rate used by
OTAA Joins:

"It is important to vary the DR. If you always choose a low DR, join
requests will take much more time on air. Join requests will also have a
much higher chance of interfering with other join attempts as well as
with regular message traffic from other devices. Conversely, if you
always use a high DR and the device trying to join the network is far
away from the LoRaWAN gateway or sitting in an RF-obstructed or null
region, the gateway may not receive a device’s join request. Given these
realities, randomly vary the DR and frequency to defend against low
signals while balancing against on-air time for join requests." [1]

I am not sure if that's a good strategy for all deployments. In my own
deployments with nodes that send Join very infrequently, I got much
better results with simply using low data rate for OTAA Joins, even on
public networks.

As a compromise, this patch adds a second optional argument to AT+JOIN
which can be used to specify the data rate for the Join. That way, the
application can choose the most appropriate data rate for its deployment
scenario.

[1] https://lora-developers.semtech.com/documentation/tech-papers-and-guides/the-book/joining-and-rejoining

Closes #43