Change first AT+JOIN parameter from retries to tries
Rather than specifying how many times to retransmit the Join in
addition to the first transmission, we let the application specify the
total number of times the Join request should be transmitted. This makes
the first parameter easier to understand. The new range for the
parameters is <1,16>, with a default value of 9.