Swap AT+JOIN parameter order
AT+JOIN takes two optional parameters. The first parameter specifies the
data rate to be used for the transmission. The second optional parameter
specifies the total number of OTAA Join transmissions.

The data rate parameter comes first because it is more likely to be
overriden by the application than the number of transmissions.

To override the number of transmissions, the application must also
specify the data rate.