Fix the channel mask format for compatibility
The AT commands AT+CHMASK and AT$CHMASK expect channel masks encoded in
a hexadecimal format. The lenght of the channel mask, and thus the the
size of its hexadecimal representation, depends on the currently active
region. Earlier firmware versions expected a long channel mask
representation with the length determined by the longest channel mask
among all compiled region. This was a bug that made AT+CHMASK
incompatible with Murata Modem.

With this patch, AT+CHMASK and AT$CHMASK expect channel mask
representations with the number of channels exactly matching the
currently active region. That is: 96 bits in CN470 (if compiled), 72
bits in US915 and AU915, and 16 bits in all other regions.

Closes #96