Add support for builds without git
When building without git or not from a git repository clone, check the
existence of files VERSION and LIB_VERSION and if they exist, extract
firmware and LoRaMac-node versions from there. If either version string
cannot be detected, report and error and refuse to build the firmware.

The files VERSION and LIB_VERSION are meant to be ignored by git. Those
files will be created by the script that creates source code tarballs.