Indicate an errors if payload is too long
When the payload supplied by the application does not fit into the
uplink message at the requested data rate, the modem internally performs
a MAC "flush", i.e., it sends an empty uplink to send any pending MAC
commands to the network server. Previous modem versions returned +OK in
this case, incorrectly indicating to the caller that the message was sent.

Since the uplink wasn't really sent, the followup +ACK/+NOACK was not
emitted.

This patch modifies the modem to correctly return +ERR=-12 if the
payload is too big to fit into the uplink message.

Closes #98