Terminate downlink payloads with CRLF
The original Murata firmware sends CRLF at the end of +RECV (downlink)
payload. This is an undocumented feature. To remain backward compatible,
send CRLF at the end of the payload as well.

Clients that use the hexadecimal payload model (AT+DFORMAT=1) can
synchronize with the incoming ATCI stream by searching for the next
occurrence of CRLF.

Closes #102