Wipe out Secure Element on region change
When the application changes region using AT+BAND=, restore all keys
stored in the Secure Element to factory defaults. This is necessary at
least for the session keys, since switching regions resets uplink and
downlink counters.

Strictly speaking, it is not necessary to wipe out the root keys on
region change, but we do that anyway to match the behavior of the
original Murata modem firmware.

Note that even though we wipe out the Secure Element, we preserve the
DevEUI value (which is stored in the Secure Element).

Closes #101