Make TCXO control pin compile-time configurable
This patch adds a new compile-time configuration option to the Makefile
that allows selecting the GPIO pin that controls TCXO_VDD power. The
Type ABZ datasheet recommends PA12 (pin 1) to be used for this purpose,
however, MKRWAN1310 boards use pin PB6. The default value is PA12. To
build a firmware version that uses PB6 use:

    make TCXO_PIN=2 release