Add AT$NVM command to store user data in NVM
The command AT$NVM makes a portion of the NVM (EEPROM) in the Type ABZ
module available to the application. The memory is available in the form
of 64 single-byte registers. The application can read and write the
values in each register individually.

Use the syntax AT$NVM 0,123 to store value 123 in NVM register 0. Use
the syntax AT$NVM 0 to read the value previously stored in NVM register
0: +OK=123.

Each NVM register can store a single byte (value 0 to 255). The values
are encoded in decimal in the ATCI. Currently, the firmware exports 64
NVM registers (0 to 63).

Closes #116