Implement LPUART1 detaching
This feature adds support for detaching the LPUART1 port (the AT command
interface port) from its GPIO. This is useful on MKRWAN1310 boards where
it can be used to give the host MCU access to the on-board flash which
shares lines with the GPIOs used by the LPUART1 port on the Murata
modem.

This feature is based on a contribution from @disk91.

Closes #119