Trigger rebuild on configuration changes
Make the build dependency tree depend on selected configuration options.
When the value of some of the option changes, e.g., because the user
provided the value for the option on the command line, trigger a rebuild
of the entire project.

The configuration is saved in the file obj/{debug,release}/config.

Closes #125