Disable interrupt wakeup on LPUART when detached
When the LPUART1 port is in a detached state, do not enable the
LPUART1 interrupt that wakes up the modem upon LPUART1 activity.
This wake up installs a stop mode lock which is only cleared upon
receiving an idle frame. However, the idle frame never arrives in
the detached state. This in turn prevented the mode from entering
the low-power stop mode.

With this commit, the modem eventually enters the low-power stop
mode when LPUART1 is detached, consuming the minimum amount of
battery power.

Closes #126