Configure debug log in MKRWAN debug builds to Segger RTT
Since USART2 cannot be used simultaneously with LPUART, configuring the
debugging logger to output to USART2 makes no sense (the output would be
lost). Since USART1 is disconnected on MKRWAN boards, we configure the
logger to output to Segger RTT.

This patch also adds few more build variants for Hardwario Chester and
the STM Discovery kit. Those variants are the same as the Hardwario
Tower variant for now.

Closes #124