Fix LoRaWAN MAC internal initialization
This commit fixes the internal initialization of the LoRaMAC-node
library. This firmware aims to be backward compatible with Murata Modem
firmware which initializes the MAC in ABP mode upon factory default. We
attempted to do the same, but the code that reconfigured the MAC into
ABP mode upon factory default was incorrect. It put the MAC into
incorrect state.

As a consequence, not all state previously saved in NVM was correctly
recovered, including a previously saved application key.

Closes #132