Enable DeviceTimeReq piggybacking
When sending a device time request via AT$DEVTIME, the application can
choose to have the request piggybacked on the next regular uplink
instead of sending an empy unconfirmed uplink immediately. This
piggybacked mode can be selected with the optional AT command argument
1, i.e., "AT$DEVTIME 1". Ommitting the argument or setting it to 0 will
trigger an immediate unconfirmed uplink with an empty payload.

Closes #136