Implement a polling mode of ATCI communication.
The modem can send a variety of asynchronous messages to the host. These
include asynchronous +EVENT notifications, +RECV messages containining
incoming (downlink) messages, +ACK and +NOACK, and +ANS. The host must
be ready to receive such messages at any time, typically seconds or tens
of seconds after the most recent AT command. An +EVENT=0,0 message
emitted after a modem reset can arrive at any time and typically
requires some response from the host, e.g., to reconfigure the modem.

Handling asynchronous modem notifications requires that the host keeps
its UART peripheral active at all times. This can be problematic on
platforms where the UART peripheral consumes significant amounts of
power while running.

This patch introduces a new ATCI command AT$ASYNC which can be used to
configure the modem's behavior with respect to asynchronous messages
sent to the host. When set to 1 (the default), the modem behaves as
before and can send asynchronous messages to the host at any time.

When set to 0 with AT$ASYNC=0, the modem switches into a polling mode of
operation. In the polling mode, the modem buffers all asynchronous
messages for the host. The buffered asynchronous messages will only be
transmitted upon receiving an AT command from the host. Upong receiving
any AT command, the modem first sends all buffered asynchronous messages
before sending the final response (+OK or +ERR) to the AT command. After
sending the final response, the modem guarantees that no asynchronous
messages will be sent until the next AT command. Thus, the host only
needs to be ready to receive between the time it sends an AT command to
the modem, and the time it receives the final response to the command.

Consider AT+JOIN as an example. In the asynchronous mode of operation,
the modem returns +OK immediately. After some time, it returns either
+EVENT=1,0 or +EVENT=1,1, depending on the outcome of the join:

| AT$ASYN=1
| +OK
|
| AT+JOIN
| +OK
|
| (...some amount of time...)
|
| +EVENT=1,1

In the polling mode, the +EVENT=1,1 message will be buffered and will
only be sent after another AT command received from the host:

| AT$ASYNC=0
| +OK
|
| AT+JOIN
| +OK
|
| (...arbitrary amount of time...)
|
| AT
| +EVENT=1,1
|
| +OK

Any AT command will trigger the sending of all buffered messages first.
If the host only wishes to poll for buffered messages, it can send the
dummy command 'AT' as shown above.

THe AT$SYNC value is backed in NVM (EEPROM) and survives modem reboots.
This is necessary because a host operating in the polling mode could
easily miss the modem's reboot indications (+EVENT=0,0).

This patch is a reinterpretation of PR #138.